use u32 instead of usize in groupby and hash indexes to reduce memory usage and improve cache performance

+++ polars/polars-core/src/series/implementations.rs
@@ -82,55 +82,51 @@ macro_rules! impl_dyn_series {
                 self.0.vec_hash(random_state)
             }
 
-            fn agg_mean(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_mean(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_mean(groups)
             }
 
-            fn agg_min(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_min(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_min(groups)
             }
 
-            fn agg_max(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_max(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_max(groups)
             }
 
-            fn agg_sum(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_sum(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_sum(groups)
             }
 
-            fn agg_first(&self, groups: &[(usize, Vec<usize>)]) -> Series {
+            fn agg_first(&self, groups: &[(u32, Vec<u32>)]) -> Series {
                 self.0.agg_first(groups)
             }
 
-            fn agg_last(&self, groups: &[(usize, Vec<usize>)]) -> Series {
+            fn agg_last(&self, groups: &[(u32, Vec<u32>)]) -> Series {
                 self.0.agg_last(groups)
             }
 
-            fn agg_std(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_std(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_std(groups)
             }
 
-            fn agg_var(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_var(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_var(groups)
             }
 
-            fn agg_n_unique(&self, groups: &[(usize, Vec<usize>)]) -> Option<UInt32Chunked> {
+            fn agg_n_unique(&self, groups: &[(u32, Vec<u32>)]) -> Option<UInt32Chunked> {
                 self.0.agg_n_unique(groups)
             }
 
-            fn agg_list(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_list(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_list(groups)
             }
 
-            fn agg_quantile(
-                &self,
-                groups: &[(usize, Vec<usize>)],
-                quantile: f64,
-            ) -> Option<Series> {
+            fn agg_quantile(&self, groups: &[(u32, Vec<u32>)], quantile: f64) -> Option<Series> {
                 self.0.agg_quantile(groups, quantile)
             }
 
-            fn agg_median(&self, groups: &[(usize, Vec<usize>)]) -> Option<Series> {
+            fn agg_median(&self, groups: &[(u32, Vec<u32>)]) -> Option<Series> {
                 self.0.agg_median(groups)
             }
 
@@ -138,7 +134,7 @@ macro_rules! impl_dyn_series {
                 &self,
                 pivot_series: &'a (dyn SeriesTrait + 'a),
                 keys: Vec<Series>,
-                groups: &[(usize, Vec<usize>)],
+                groups: &[(u32, Vec<u32>)],
                 agg_type: PivotAgg,
             ) -> Result<DataFrame> {
                 self.0.pivot(pivot_series, keys, groups, agg_type)
@@ -148,23 +144,23 @@ macro_rules! impl_dyn_series {
                 &self,
                 pivot_series: &'a (dyn SeriesTrait + 'a),
                 keys: Vec<Series>,
-                groups: &[(usize, Vec<usize>)],
+                groups: &[(u32, Vec<u32>)],
             ) -> Result<DataFrame> {
                 self.0.pivot_count(pivot_series, keys, groups)
             }
-            fn hash_join_inner(&self, other: &Series) -> Vec<(usize, usize)> {
+            fn hash_join_inner(&self, other: &Series) -> Vec<(u32, u32)> {
                 HashJoin::hash_join_inner(&self.0, other.as_ref().as_ref())
             }
-            fn hash_join_left(&self, other: &Series) -> Vec<(usize, Option<usize>)> {
+            fn hash_join_left(&self, other: &Series) -> Vec<(u32, Option<u32>)> {
                 HashJoin::hash_join_left(&self.0, other.as_ref().as_ref())
             }
-            fn hash_join_outer(&self, other: &Series) -> Vec<(Option<usize>, Option<usize>)> {
+            fn hash_join_outer(&self, other: &Series) -> Vec<(Option<u32>, Option<u32>)> {
                 HashJoin::hash_join_outer(&self.0, other.as_ref().as_ref())
             }
             fn zip_outer_join_column(
                 &self,
                 right_column: &Series,
-                opt_join_tuples: &[(Option<usize>, Option<usize>)],
+                opt_join_tuples: &[(Option<u32>, Option<u32>)],
             ) -> Series {
                 ZipOuterJoinColumn::zip_outer_join_column(&self.0, right_column, opt_join_tuples)
             }
@@ -183,7 +179,7 @@ macro_rules! impl_dyn_series {
             fn remainder(&self, rhs: &Series) -> Result<Series> {
                 NumOpsDispatch::remainder(&self.0, rhs)
             }
-            fn group_tuples(&self, multithreaded: bool) -> Vec<(usize, Vec<usize>)> {
+            fn group_tuples(&self, multithreaded: bool) -> Vec<(u32, Vec<u32>)> {
                 IntoGroupTuples::group_tuples(&self.0, multithreaded)
             }
         }
@@ -680,7 +676,7 @@ macro_rules! impl_dyn_series {
                 ChunkUnique::n_unique(&self.0)
             }
 
-            fn arg_unique(&self) -> Result<Vec<usize>> {
+            fn arg_unique(&self) -> Result<Vec<u32>> {
                 ChunkUnique::arg_unique(&self.0)
             }
 
@@ -1086,7 +1082,7 @@ where
         ChunkUnique::n_unique(&self.0)
     }
 
-    fn arg_unique(&self) -> Result<Vec<usize>> {
+    fn arg_unique(&self) -> Result<Vec<u32>> {
         ChunkUnique::arg_unique(&self.0)
     }
